pick changes from feature/v3.7.x/jetson_uefi_rebased branch
