+++ tests/test_boot_control/test_jetson_cboot.py
@@ -25,12 +25,13 @@ from typing import Any
 import pytest
 
 from otaclient.app.boot_control import _jetson_cboot
-from otaclient.app.boot_control._jetson_cboot import (
+from otaclient.app.boot_control._jetson_cboot import _CBootControl
+from otaclient.app.boot_control._jetson_common import (
     BSPVersion,
     FirmwareBSPVersion,
     SlotID,
-    _CBootControl,
     parse_bsp_version,
+    update_extlinux_cfg,
 )
 
 logger = logging.getLogger(__name__)
@@ -140,4 +141,4 @@ def test_parse_bsp_version(_in: str, expected: BSPVersion):
 def test_update_extlinux_conf(_template_f: Path, _updated_f: Path, partuuid: str):
     _in = (TEST_DIR / _template_f).read_text()
     _expected = (TEST_DIR / _updated_f).read_text()
-    assert _CBootControl.update_extlinux_cfg(_in, partuuid) == _expected
+    assert update_extlinux_cfg(_in, partuuid) == _expected
